+++ dubbo-config/dubbo-config-spring/src/main/resources/META-INF/dubbo.xsd
@@ -1210,6 +1210,7 @@
     <xsd:complexType name="baggageType">
         <xsd:all>
             <xsd:element ref="remoteFields" minOccurs="0"/>
+            <xsd:element ref="correlation" minOccurs="0"/>
         </xsd:all>
         <xsd:attribute name="enabled" type="xsd:boolean">
             <xsd:annotation>
@@ -1236,6 +1237,33 @@
                 <xsd:attribute name="name" type="xsd:string"/>
             </xsd:extension>
         </xsd:simpleContent>
+    </xsd:complexType>
+
+    <xsd:complexType name="tracingExporterType">
+        <xsd:all>
+            <xsd:element ref="zipkin-config" minOccurs="0"/>
+        </xsd:all>
+    </xsd:complexType>
+
+    <xsd:complexType name="zipkinConfigType">
+        <xsd:attribute name="endpoint" type="xsd:string">
+            <xsd:annotation>
+                <xsd:documentation>
+                    <![CDATA[ The url of zipkin server. ]]></xsd:documentation>
+            </xsd:annotation>
+        </xsd:attribute>
+        <xsd:attribute name="connectTimeout" type="xsd:duration">
+            <xsd:annotation>
+                <xsd:documentation>
+                    <![CDATA[ Connection timeout for requests to Zipkin. ]]></xsd:documentation>
+            </xsd:annotation>
+        </xsd:attribute>
+        <xsd:attribute name="readTimeout" type="xsd:duration">
+            <xsd:annotation>
+                <xsd:documentation>
+                    <![CDATA[ Read timeout for requests to Zipkin. ]]></xsd:documentation>
+            </xsd:annotation>
+        </xsd:attribute>
     </xsd:complexType>
 
     <xsd:complexType name="methodType">
@@ -2185,6 +2213,20 @@
             <xsd:documentation>
                 <![CDATA[ List of fields that should be correlated with the logging context. That
                 means that these fields would end up as key-value pairs in e.g. MDC. ]]></xsd:documentation>
+        </xsd:annotation>
+    </xsd:element>
+
+    <xsd:element name="tracing-exporter" type="tracingExporterType">
+        <xsd:annotation>
+            <xsd:documentation>
+                <![CDATA[ Exporter of tracing. ]]></xsd:documentation>
+        </xsd:annotation>
+    </xsd:element>
+
+    <xsd:element name="zipkin-config" type="zipkinConfigType">
+        <xsd:annotation>
+            <xsd:documentation>
+                <![CDATA[ Config of zipkin exporter. ]]></xsd:documentation>
         </xsd:annotation>
     </xsd:element>
 </xsd:schema>
