+++ dubbo-demo/dubbo-demo-spring-boot/dubbo-demo-spring-boot-consumer/pom.xml
@@ -125,7 +125,7 @@
 
         <dependency>
             <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-spring-boot-observability-starter</artifactId>
+            <artifactId>dubbo-spring-boot-tracing-otel-zipkin-starter</artifactId>
         </dependency>
 
         <dependency>
