+++ dubbo-demo/dubbo-demo-spring-boot/dubbo-demo-spring-boot-provider/pom.xml
@@ -120,7 +120,7 @@
         <!-- Observability -->
         <dependency>
             <groupId>org.apache.dubbo</groupId>
-            <artifactId>dubbo-spring-boot-observability-starter</artifactId>
+            <artifactId>dubbo-spring-boot-tracing-otel-zipkin-starter</artifactId>
         </dependency>
 
         <dependency>
