--- dubbo-spring-boot/dubbo-spring-boot-observability-starter/src/main/resources/META-INF/spring.factories
@@ -1,6 +1,0 @@
-org.springframework.boot.autoconfigure.EnableAutoConfiguration=\
-org.apache.dubbo.spring.boot.observability.autoconfigure.otel.OpenTelemetryAutoConfiguration,\
-org.apache.dubbo.spring.boot.observability.autoconfigure.DubboMicrometerTracingAutoConfiguration,\
-org.apache.dubbo.spring.boot.observability.autoconfigure.DubboObservationAutoConfiguration,\
-org.apache.dubbo.spring.boot.observability.autoconfigure.brave.BraveAutoConfiguration
-
