--- dubbo-spring-boot/dubbo-spring-boot-observability-starter/src/main/resources/META-INF/spring/org.springframework.boot.autoconfigure.AutoConfiguration.imports
@@ -1,4 +1,0 @@
-org.apache.dubbo.spring.boot.observability.autoconfigure.otel.OpenTelemetryAutoConfiguration
-org.apache.dubbo.spring.boot.observability.autoconfigure.DubboMicrometerTracingAutoConfiguration
-org.apache.dubbo.spring.boot.observability.autoconfigure.DubboObservationAutoConfiguration
-org.apache.dubbo.spring.boot.observability.autoconfigure.brave.BraveAutoConfiguration
